fix(security): resolve Edge Runtime compatibility issues
- Add dynamic export to /api/auth/permissions to fix static generation error
- Move useSecureLogout hook to separate client-side file to avoid next-auth/react Edge Runtime conflicts
- Separate client-only functionality from Edge Runtime compatible utilities
- Fix build errors related to BroadcastChannel Node.js API usage
